CTSType is generated but...
The TSType cache is NOT good. Some recursive type resolution triggers
reentrant calls to the "by Object" resolution.
This is due to named record default value existence computation... but
if we have no knowledge of the default values we cannot compute
anonymous records type name.

The thing is: I have currently no idea on how this can be solved except
introducing a TSTypePromise and defer almost any process thanks
to a trampoline... OR binding the "a default value exist" to the C#
behavior.
This may be the best option... Let's sleep on that one.