Stop setting the type after nosj call.
This causes troubles between boxed object and
primitive types. Let the constructor of what is a
class set it but don't transform primitives.

Another (best) option would be to skip primitives
(because of the troubles)
and classes (because already done) but set the SymCTS for
collections (array, set, map). This would also work for
value tuples as arrays. Value tuples as objects would need
to be detected.

Eventually calling it on typeof == 'object' should do the job:
for already tagged class instances this is a no-op.