WIP. Question: Should TypeScriptRoot be a IResourceContainer?
More generally, generated code would be stored in an (intermediate) resource container.
That would allow transformers to apply to generated code.
This means that a ResourceSpace is bound to (at least) one CodeGenResourceContainer.

The current pattern is that any ResPackage has an available CodeGenResourceContainer.
The idea is that code can add generated resources to each ResPackage:
those CodeGenResourceContainer are CLOSED by ResourceSpaceDataBuilder.Build
before producing the final ResourceSpaceData: these code generated resources
(if any) are available to the ResourceFile/FolderHandlers.
This is the key to support (in the future) "meta code generator" ("Patterns"
and other automatic complex constructs).

It is obvious that the TS code generator happens before the potential transformers.
Moreovoer, the "types" generated are unknow to the resource layer. For code
generation to be included in the transformer space, it seems that code generators
should actually be "ResPackage generators". In a way, it's what the TypeScriptRoot
offers. Code generators create TypeScriptFolders and TypeScriptFiles in a
TypeScriptRoot where some of the TypeScriptFiles are bound to a C# type
and others are not, just like the new ResPackage. TypeScriptRoot does
more than holding the TypeScript folder/file structure as it maintains
the js librairies registration, generates the relative paths, etc.
It cannot be replaced (unification should rely on it) but its current API
(based on folder objects) may be changed to rely more on paths
like the resource containers do (but this is a detail).
A huge difference between ResPackages and TypeScriptRoot folders/files
is that there is no (at least no need for) dependencies in the topological
sense of the ResPackages between the TypeScriptRoot items. Dependencies
are already managed by TypeScript import statements (and
co-dependencies are often required) => A code generated ResPackage is
unique and contains the bunch of generated resources.

Can we consider that a ResourceSpace depends on a single CodeGenerated
ResPackage?
We'd need to implement TypeScriptRoot : IResourceContainer (that shouldn't
be hard). A harder issue seems to transparently consider the CodeGenerated
resources in the same namespace as the "regular" resources.
A CodeGenerated resource cannot "overwrite/replace" an actual resource file
and vice-versa. But... wait... Resource namespace management is more subtle
than that. A resource name is unique in its ResPackage (by design), not
necessarily in the final target namespace (this is the motivation of
the ResPackage.DefaultTargetPath).
To understand this, it is enough to compare 2 kind of resources:
- The Assets: they are here to be ultimately produced. The same
resource name can exist in more than one ResPackage and can be
mapped to the same final target name. This final target name depends
on the DefaultTargetPath or on the "targetPath" property or by an
explicit "mappings": [{ "name": "target" }] of the "assets.jsonc" file.
It is up to the LoadAssets extension methods to resolve this and then
to the FinalResourceAssetSet to combine multiple asset sets in a
final set and detect any conflicting mapping of mutiple source to
the same final target name.
- Locales: they are here to define translation dictionnaries. They have
absolutely nothing to do with a "final target name" concept. The
LoadLocales extension methods has no defaultTargetPath, it only
computes LocaleCultureSet that are combined by FinalLocaleCultureSet
into a tree of translation dictionaries tat can be saved as one
or more files or can even be transformed in a set of FrozenDictionary
initialization code.

The lesson here is that resources are nothing by themselves. It
is how they are handled that matters: some of them are like
"configuration" files that are only read, others are here
to be copied in a target container (file system), maybe after
a sequence of transformations (that may alter their theorethical
"target name").

The question is then: how the TypeScript generator that can
play its "pure code generation" partition AND its "resource-based
management" recorded music at the same time?