Questions....
Should ResourceFolder.FolderName ends with the / or \?
It may not as this is not the "standard" but above all, this
requires a string alloc while working with the file system and
with resource container thar are "virtual" like the TypeScriptRoot
that work with NormalizedPath. This breaks homogeneity...
For the AssemblyResourceContainer, by assuming the IResourceContainer.MaxNameLength
path manipulations can use stack alloc.

We should also find a way to skip the ResourceFolder/Locator guards that can
be costly. Static UnsafeCreate in addition to the public guarded ctors?

Is the CodeStoreResources a good idea? Another approach would be to
consider that Store resources that are "used"/"manipulated" by code are
"removed" from the Store (hidden) and the code is responsible to create
them (or not) in the head "<Code>" container. This transfers a part of the
complexity to the code generator, if it wants/needs to play with the Store
resources and frees the "Live" of storing/handling today's CodeStoreResources.Code
immutable content.
"Live" would only have to handle Store resources (the "<Code>" is just a
non-local container like the others).