Added Bind support... but...
Bind seems hard to maintain. Parsing & Binding
are too close, they overlap (too much).
When head tokens of a SourceSpan are touched,
the SourceSpan may need to be mutated into
another SourceSpan type and this can be decided
by the parent construct... and what if a sub span type
conditions the parent span type (the bottom-up Parse
naturally allows this behavior even if this is weird)?
Binding (correctly) seems too complex to be easy to maintain.

It seems that if we accept the existence of weird languages
in which sub-constructs have impacts on higher constructs, there
is no other way than to reparse (bottom-up).
The goal is now to reparse "as locally as possible". This can
only be decide by the language. A KISS solution may be that
SourceSpana has a simple dirty flag and a OnTokenChange()
called by the editor. This OnTokenChange may call
_parent.OnTokenChange depending on the change... This would
enable the SourceSpan to decide where the reparse points are
or (if possible) to immediately account for the changes...

The good things here is that implementations can be
basic and simple and optimized but complex when required
(that is one design goal of this library).

=> Bye bye Bindings!