Trying Editor facade approach.
This facade will expose the Token mutation API.
Closing the Editor gives us the "reparse now" point.
This replaces/generalizes the current TokenScope:
SourceTokens can only be obtained from it. This
restricts the SourceToken enumerator to be a single
one: any code modification can only be done on the basis
of the SourceTokens and the root enumerator has no reason
to be acquired more than once.