Wrong:! (again)
The SpanMatcher parsing cannot be handled by the TransformStatementAnalyzer.
It must be the TargetAnalyzer. This means that to transform transformer,
TransformStatementAnalyzer must be ITargetAnalyzer.

The factorized code of ParseSpanSpec & ParsePattern must be moved to the
ITargetAnalyzer... which means this should be an abstract base (DIM is
not a good idea).
Will first remove the .language in the SpanMatcher: this is currently more
that overkill (was meant to match inside patterns: transforming patterns
will require modifying the whole string).

But even with this removal, transforming a transformer must be able to
"see" specific tokens that a transform language use. Transfomer of
Transformers (Transformer²) must be dynamically created on the targeted
language...

Transformer² is necessarily an instance of the "base transform language".
The "transformer" language doesn't exist. Only "ts.t" and "less.t" exist.
And then "ts.t.t" are on-demand instances of the base transform language.