Working... but...
...when there are matches: empty ranges must be detected.

Possible existence of empty ranges either forces
to check their existence EVERYWHERE or to check this
in the final "Apply" and then lose the precise "location"
of the error.

Currently, there is only 2 SourceToken generators: the TokenSpanFilter
that never returns an empty range and the SingleSpanTypeFilter
with its DeepestSpanCollector. This last one CAN generate empty
ranges when its SourceSpan has been detached by a editor.Replace.

If we keep it this way, not adding union/intersection and other stuffs
that will REALLY change the architecture (the current design is focused
on a single chain of filters that enables the tracking of the
modifications) AND if the after/before/between implementation can
also avoid generating empty ranges, then we can prune the
DeepestSpanCollector List<List<SourceSpan>> of its empty ranges
BEFORE returning its enumerator, then we're done.

So now the after/before/between (RangeLocation) must be implemented
and if it can guaranty to not generate empty ranges (under the condition
that it's inner doesn't have empty range) then we are good: it will be
up to the Cardinality operators (and only them) to do their work.

Note: the current LinkedTokenFilterBuilderContext.Error must be
refactored for 2 reasons: the code is messy - an intermediate
TokenFilteringDiagnostic that can concretize the filter once for all
is welcome - and the dump must be deferred (once all the enumerators
are released).