Introduced ResourceTypeScriptFile.IsPublishedResource.
Resources "moved" to the Code Generated Container can now
be left in their original container (no more hidden) or hidden
and "published as code".
This is used by the [TypeScriptFile] to allow change of the
resource target path (that cannot be done at the container level).

This has currently one drawback: Code Generated Resources are not
transformable BUT there's no real reason for this (I'm tired, will
investigate this tomorrow).