Heavy refactoring...
The ITokenFilterAnchoredOperator model is stupid.
The {xxx} where "pattern" IS a simple where filter of the matches.
No match should be perfectly valid: this would greatly simplify
the horrible current error management. We'll introduce a
required/optional operator that will be triggered by a trailing '?'
(coalescing operator). Default has to be required (just like NRT).
A "single" cardinality for instance is not an error on no match, it only
error if there's more than one match.
Same for any other cardinality operator, even if this seems weird:
"in each first 5 out 7 ..." must error on card != 7 but not 0. The if
it is "in each first 5 out 7 ... ?" there's no error when 0. Only if the
'?' doesn't appear do we error on 0.

The '?' applies to the InScope (that is a "sub statement") only.
For instance the "replace ... with "..." cannot use
"replace ...? with "...". To express this, it is enough to write:
in ...?
   replace * with "...";