Failure...
This implements the IResPackageDescriptorResolver.
But this is not enough. Type mapping must be expressed, exposed.
The current PackageIndex may be the SingleMapping map.
To cover other scenarii (real object), multiple mappings should also
be supported AND navigation in the type model must be easier.

One idea is to rely on the ResPackageDescriptor to capture mappings
thanks to ResPackageDescriptor.AddMapping( Type ), AddMapping( string ),
AddMultipleMapping( Type ), AddMultipleMapping( string ).

=> One of the goal is to be able to express a system purely with string
(type.FullName being the projection of the type).

Navigating in the type mappings will complicate the ResPackage API.
But this seems to be required. A dedicated TypeModel, or may be
a more powerful PackageIndex than a dictionnary can do this. This
enhanced PackageIndex would map strings and C# types to
ResPackageDescriptor with a support of Single and Multiple mappings.

Another aspect is the association to a "Type descriptor", a "Type Driver"
or something similar that "implements" the ResPackageDescriptor (i.e
that implements its Type when a ResPackageDescriptor is defined by
a Type). This "Driver" is perfect to support the package textual
description (package.xml contains the driver type and its configuration).

This Type driver corresponds to the "primary attribute" that we're using
in CK-DB as well as in CK-TypeScript through the pattern:

[EntityType]
class X : EntityType {}

This pattern proved to be easy to use and seems to capture all
our needs. In CK-TypeScript the notion of "extensions"
(the awful TypeScriptGroupOrPackageAttributeImplExtension)
works well and is esay to understand. This could be generalized.

Perhaps the next idea is to say that a ResPackageDescriptor is
not bound to a Type, but to a ResPackageDriver (that MAY be
a "Type born" one). Today, the association from a Type to its
"driver" must be done externally (TypeScriptContext or the
StObjMap for instance). With this, this key association will
be done through the PackageIndex of the ResSpace that
maps Type to ResPackage and ResPackage that maps
directly to its Driver.

This approach put the ResSpace at the heart of the architecture:
a ResSpace is a mini-StObjMap. By selecting a set of IRealObject
and IAutoService, will this be enough to reproduce the current
kernel?
Is the whole AutoDI a simple plugin of the ResSpace?
If we follow this road, what is the purpose of CK-ReaDI? It seems
that lifetime control is going back to the ResSpace AND to the
ResPackageDriver. Eventually, an Engine Aspect is in charge
of creating one or more ResSpace and install them... and
is the orchestrator that have full control of the execution.

Today a CKomposable architecture is not data driven. This is
the goal of the TODO package.xml. This ResPackageDriver
would allow a data representation of an architecture. If
a ResPackageDriver is "serializable", or "exportable" we
may be able to invent new ways, new tools to develop
packages. (The Live is key here: may be some ResPackageDriver
could have a Live counterpart...)

It may be time to take a breath and rething the whole System
top-down. From Engine, Aspects and ResSpace. Every concerns
should be addressed. For instance the "documentation system"
that would be able to scaffold a central doc from "Res/Doc" should
be able to work with any kind of package.... The ResPackageDriver
is useless for its job (a ResourceSpaceFolderHandler should be enough).
The Engine Aspect is the key here: it is the one that must setup the
ResSpaces. It must be able to NOT use the ResPackageDriver of a type
(by associating a default one to the ResPackageDescriptor?).

The ResSpace model must remain a "tool", a passive component that is
configured "from the outside" by the Engine Aspects.