Better! Optionality is... optional.
The IOptionalResourceGroupAttribute interface enables a
"PrimaryKindAttribute" (should be better named) that defines
the "kind" or IResourceGroup/Package (ie. TypeScriptGroupAttribute and
TypeScriptPackageAttribute) to support optionality... or not.

Should there be a similar IAbstractResourceGroupAttribute that drives
a IsAbstract property (overridable just like optionality when
registering?).

This IsAbstract would imply IsOptional... An abstract must have
non optional registered specializations to become non-optional
(or, may be better, if it is not optional and there is no
specializations, then it's an error).

To handle this we would need a
ResPackageDescriptor? Generalization { get; set; }
property that will be used when initializing the graph.
Note: Generalization property enables both top-down and bottom-up
ResPackageDescriptor initializations.

Generalization implies a Requires (from the specialization to its
Generalization) but not a SingleMapping at the ResPackageDescriptor
level: we don't want relationships (like the Requires) to be projected
to the specialization, they must stick to the Generalization...
The ResPackageDescriptor's SingleMapping is only a part of the final
DI SingleMapping: the DI SingleMapping also includes the Generalization
to Specialization WHEN the Generalization is in a way or another marked
as a "Single" concept.

Would there be a third similar mechanism for Single vs. Multiple
concepts?
Whether a ResPackageDescriptor is Single or Multiple seems an orthogonal
concept regarding Concrete/Abstract.

Is a ResPackageDescriptor necessarily Single xor Multiple (i.e
is there a "None" mapping)?
My intuition is that:
- There is no "None" mapping. A concept is either a
   Single or a Multiple one.
- And that the Single/Multiple and Concrete/Abstract are
  not that orthogonal :-(
Idea: Should the IsAbstract be expanded to Abstract/Concrete/Sealed?
In this scenario, Sealed naturally implies not Single and not Multiple
(ie. the "None" mapping).

So... "Sealed" prevents any specialization to be defined and that seems
a good point. Putting IsOptional as independent (that seems correct),
the valid Abstractness/Cardinality could be:

* Sealed: Basic concrete non specializable concept.
* ConcreteSingle: Specializable concept. The UserTable that can be
                  specialized by ZoneUserTable.
* AbstractSingle: rather standard combination
	          (like a [CKTypeDefiner] on a IRealObject)
* AbstractMutiple: another standard combination, the current [IsMultiple]
	           attribute (currently only on interfaces) or a
	           [CKTypeDefiner] on a IPoco (an Abstract IPoco).
* ConcreteMutiple: This one is problematic. What does it mean?
                   It seems that this combination is useless.
		   The only scenario I can imagine is a
		default implementation that could be
		"co-implemented" by other ones... It is
		the far more easy to consider an abstraction with
		a default implementation that can be replaced
		by another one (the [ReplaceService] in the AutomaticDI).

It seems that Multiple => Abstract and this is exactly the current state:
[IsMultiple] can only decorate interfaces (and it SHOULD be able to also
decorate abstract classes - the current kernel forbids this).

Eventually, we have:
- The IOptionalResourceGroupAttribute that exposes IsOptional.
- A IKindResourceGroupAttribute (?) that exposes a
   Sealed/ConcreteSingle/AbstractSingle/AbstractMutiple enum.

This "Kind" is the Abstractness/Cardinality specifier ("ACSpecifier",
"ACSM", "SMAC"? I like the last one :-)).
What's the AC default?
Sealed seems to be the best choice as it is the most restrictive one.

Should Optionality/Abstractness/Cardinality be handled by a single interface?
When multiple attributes decorate a Type, IsOptional can be combined (false wins,
required is the default), but how
Sealed/ConcreteSingle/AbstractSingle/AbstractMutiple can be combined?
There's no obvious ansxer here. But we can choose to forbid such multiple (ambiguous)
decorations and require at most one rule. This may be a goog thing: the "PrimaryKindAttribute"
has this responsibility. A constructor overload can accept the bool isOptional and the enum
if this makes sense for the "kind" (for rich model) or specialized attributes can be created
to ease the discovery or better communicate the intent like "OptionalTypeScriptPackageAttribute" and
"TypeScriptPackageAttribute".

So... Only one interface on at most one attribute that is the "PrimaryKindAttribute"?
Then it is the:

enum ResourceGroupKind { Sealed, ConcreteSingle, AbstractSingle, AbstractMutiple }

interface IResourceGroupKindAttribute
{
  bool IsOptional { get; set }
  ResourceGroupKind Kind { get; set }
}

And Attributes can use explicit implementations to hide the properties when needed.

Should the Abstractness/Cardinality be overridable at registration just like the
IsOptional (with a nullable parameter). Definitly yes: this enables external
computation of this aspect if needed in some scenario that can use unknown
artefacts.
Note: This external computation approach for instance by using other
      attributes on the type, the name of the base class of the decorated type like
      [TypeScriptPackage]class MyPackage : OptionalTypeScriptPackage {}, etc.
      may not be a great idea when, if the ResourceSpace becomes the new kernel,
      heterogeneous "PrimaryKind" types must be registered in some resource spaces to handle
      transversal aspects.

Looks good...